weather_api/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
python-decouple==3.8
idna==3.10
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
